+++ stac_fastapi/extensions/stac_fastapi/extensions/core/pagination/token_pagination.py
@@ -1,6 +1,6 @@
 """Token pagination API extension."""
 
-from typing import List, Optional
+from typing import Optional
 
 import attr
 from fastapi import FastAPI
@@ -22,7 +22,7 @@ class TokenPaginationExtension(ApiExtension):
     GET = GETTokenPagination
     POST = POSTTokenPagination
 
-    conformance_classes: List[str] = attr.ib(factory=list)
+    conformance_classes: list[str] = attr.ib(factory=list)
     schema_href: Optional[str] = attr.ib(default=None)
 
     def register(self, app: FastAPI) -> None: