+++ stac_fastapi/extensions/stac_fastapi/extensions/core/sort/sort.py
@@ -1,5 +1,6 @@
 """Sort extension."""
-from typing import List, Optional
+
+from typing import Optional
 
 import attr
 from fastapi import FastAPI
@@ -21,7 +22,7 @@ class SortExtension(ApiExtension):
     GET = SortExtensionGetRequest
     POST = SortExtensionPostRequest
 
-    conformance_classes: List[str] = attr.ib(
+    conformance_classes: list[str] = attr.ib(
         factory=lambda: ["https://api.stacspec.org/v1.0.0/item-search#sort"]
     )
     schema_href: Optional[str] = attr.ib(default=None)