+++ stac_fastapi/types/stac_fastapi/types/core.py
@@ -1,25 +1,25 @@
 """Base clients."""
+
 import abc
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Optional, Union
 from urllib.parse import urljoin
 
 import attr
 from fastapi import Request
+from stac_pydantic import Collection, Item, ItemCollection, api
 from stac_pydantic.links import Relations
 from stac_pydantic.shared import MimeTypes
 from stac_pydantic.version import STAC_VERSION
 from starlette.responses import Response
 
-from stac_fastapi.types import stac as stac_types
 from stac_fastapi.types.conformance import BASE_CONFORMANCE_CLASSES
 from stac_fastapi.types.extension import ApiExtension
 from stac_fastapi.types.requests import get_base_url
 from stac_fastapi.types.search import BaseSearchPostRequest
-from stac_fastapi.types.stac import Conformance
 
 NumType = Union[float, int]
-StacType = Dict[str, Any]
+StacType = dict[str, Any]
 
 
 @attr.s  # type:ignore
@@ -30,9 +30,9 @@ class BaseTransactionsClient(abc.ABC):
     def create_item(
         self,
         collection_id: str,
-        item: Union[stac_types.Item, stac_types.ItemCollection],
+        item: Union[Item, ItemCollection],
         **kwargs,
-    ) -> Optional[Union[stac_types.Item, Response, None]]:
+    ) -> Optional[Union[Item, Response, None]]:
         """Create a new item.
 
         Called with `POST /collections/{collection_id}/items`.
@@ -48,8 +48,8 @@ class BaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     def update_item(
-        self, collection_id: str, item_id: str, item: stac_types.Item, **kwargs
-    ) -> Optional[Union[stac_types.Item, Response]]:
+        self, collection_id: str, item_id: str, item: Item, **kwargs
+    ) -> Optional[Union[Item, Response]]:
         """Perform a complete update on an existing item.
 
         Called with `PUT /collections/{collection_id}/items`. It is expected
@@ -69,7 +69,7 @@ class BaseTransactionsClient(abc.ABC):
     @abc.abstractmethod
     def delete_item(
         self, item_id: str, collection_id: str, **kwargs
-    ) -> Optional[Union[stac_types.Item, Response]]:
+    ) -> Optional[Union[Item, Response]]:
         """Delete an item from a collection.
 
         Called with `DELETE /collections/{collection_id}/items/{item_id}`
@@ -85,8 +85,8 @@ class BaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     def create_collection(
-        self, collection: stac_types.Collection, **kwargs
-    ) -> Optional[Union[stac_types.Collection, Response]]:
+        self, collection: Collection, **kwargs
+    ) -> Optional[Union[Collection, Response]]:
         """Create a new collection.
 
         Called with `POST /collections`.
@@ -101,8 +101,8 @@ class BaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     def update_collection(
-        self, collection: stac_types.Collection, **kwargs
-    ) -> Optional[Union[stac_types.Collection, Response]]:
+        self, collection: Collection, **kwargs
+    ) -> Optional[Union[Collection, Response]]:
         """Perform a complete update on an existing collection.
 
         Called with `PUT /collections`. It is expected that this item already
@@ -122,7 +122,7 @@ class BaseTransactionsClient(abc.ABC):
     @abc.abstractmethod
     def delete_collection(
         self, collection_id: str, **kwargs
-    ) -> Optional[Union[stac_types.Collection, Response]]:
+    ) -> Optional[Union[Collection, Response]]:
         """Delete a collection.
 
         Called with `DELETE /collections/{collection_id}`
@@ -144,9 +144,9 @@ class AsyncBaseTransactionsClient(abc.ABC):
     async def create_item(
         self,
         collection_id: str,
-        item: Union[stac_types.Item, stac_types.ItemCollection],
+        item: Union[Item, ItemCollection],
         **kwargs,
-    ) -> Optional[Union[stac_types.Item, Response, None]]:
+    ) -> Optional[Union[Item, Response, None]]:
         """Create a new item.
 
         Called with `POST /collections/{collection_id}/items`.
@@ -162,8 +162,8 @@ class AsyncBaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     async def update_item(
-        self, collection_id: str, item_id: str, item: stac_types.Item, **kwargs
-    ) -> Optional[Union[stac_types.Item, Response]]:
+        self, collection_id: str, item_id: str, item: Item, **kwargs
+    ) -> Optional[Union[Item, Response]]:
         """Perform a complete update on an existing item.
 
         Called with `PUT /collections/{collection_id}/items`. It is expected
@@ -182,7 +182,7 @@ class AsyncBaseTransactionsClient(abc.ABC):
     @abc.abstractmethod
     async def delete_item(
         self, item_id: str, collection_id: str, **kwargs
-    ) -> Optional[Union[stac_types.Item, Response]]:
+    ) -> Optional[Union[Item, Response]]:
         """Delete an item from a collection.
 
         Called with `DELETE /collections/{collection_id}/items/{item_id}`
@@ -198,8 +198,8 @@ class AsyncBaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     async def create_collection(
-        self, collection: stac_types.Collection, **kwargs
-    ) -> Optional[Union[stac_types.Collection, Response]]:
+        self, collection: Collection, **kwargs
+    ) -> Optional[Union[Collection, Response]]:
         """Create a new collection.
 
         Called with `POST /collections`.
@@ -214,8 +214,8 @@ class AsyncBaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     async def update_collection(
-        self, collection: stac_types.Collection, **kwargs
-    ) -> Optional[Union[stac_types.Collection, Response]]:
+        self, collection: Collection, **kwargs
+    ) -> Optional[Union[Collection, Response]]:
         """Perform a complete update on an existing collection.
 
         Called with `PUT /collections`. It is expected that this item already
@@ -234,7 +234,7 @@ class AsyncBaseTransactionsClient(abc.ABC):
     @abc.abstractmethod
     async def delete_collection(
         self, collection_id: str, **kwargs
-    ) -> Optional[Union[stac_types.Collection, Response]]:
+    ) -> Optional[Union[Collection, Response]]:
         """Delete a collection.
 
         Called with `DELETE /collections/{collection_id}`
@@ -260,10 +260,10 @@ class LandingPageMixin(abc.ABC):
     def _landing_page(
         self,
         base_url: str,
-        conformance_classes: List[str],
-        extension_schemas: List[str],
-    ) -> stac_types.LandingPage:
-        landing_page = stac_types.LandingPage(
+        conformance_classes: list[str],
+        extension_schemas: list[str],
+    ) -> api.LandingPage:
+        landing_page = api.LandingPage(
             type="Catalog",
             id=self.landing_page_id,
             title=self.title,
@@ -320,13 +320,13 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
         extensions: list of registered api extensions.
     """
 
-    base_conformance_classes: List[str] = attr.ib(
+    base_conformance_classes: list[str] = attr.ib(
         factory=lambda: BASE_CONFORMANCE_CLASSES
     )
-    extensions: List[ApiExtension] = attr.ib(default=attr.Factory(list))
+    extensions: list[ApiExtension] = attr.ib(default=attr.Factory(list))
     post_request_model = attr.ib(default=BaseSearchPostRequest)
 
-    def conformance_classes(self) -> List[str]:
+    def conformance_classes(self) -> list[str]:
         """Generate conformance classes by adding extension conformance to base
         conformance classes."""
         base_conformance_classes = self.base_conformance_classes.copy()
@@ -351,7 +351,7 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
 
         return base_conformance
 
-    def landing_page(self, **kwargs) -> stac_types.LandingPage:
+    def landing_page(self, **kwargs) -> api.LandingPage:
         """Landing page.
 
         Called with `GET /`.
@@ -405,7 +405,7 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
 
         return landing_page
 
-    def conformance(self, **kwargs) -> stac_types.Conformance:
+    def conformance(self, **kwargs) -> api.ConformanceClasses:
         """Conformance classes.
 
         Called with `GET /conformance`.
@@ -413,12 +413,12 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
         Returns:
             Conformance classes which the server conforms to.
         """
-        return Conformance(conformsTo=self.conformance_classes())
+        return api.ConformanceClasses(conformsTo=self.conformance_classes())
 
     @abc.abstractmethod
     def post_search(
         self, search_request: BaseSearchPostRequest, **kwargs
-    ) -> stac_types.ItemCollection:
+    ) -> api.ItemCollection:
         """Cross catalog search (POST).
 
         Called with `POST /search`.
@@ -434,18 +434,18 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
     @abc.abstractmethod
     def get_search(
         self,
-        collections: Optional[List[str]] = None,
-        ids: Optional[List[str]] = None,
-        bbox: Optional[List[NumType]] = None,
+        collections: Optional[list[str]] = None,
+        ids: Optional[list[str]] = None,
+        bbox: Optional[list[NumType]] = None,
         datetime: Optional[Union[str, datetime]] = None,
         limit: Optional[int] = 10,
         query: Optional[str] = None,
         token: Optional[str] = None,
-        fields: Optional[List[str]] = None,
+        fields: Optional[list[str]] = None,
         sortby: Optional[str] = None,
         intersects: Optional[str] = None,
         **kwargs,
-    ) -> stac_types.ItemCollection:
+    ) -> api.ItemCollection:
         """Cross catalog search (GET).
 
         Called with `GET /search`.
@@ -456,7 +456,7 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
         ...
 
     @abc.abstractmethod
-    def get_item(self, item_id: str, collection_id: str, **kwargs) -> stac_types.Item:
+    def get_item(self, item_id: str, collection_id: str, **kwargs) -> api.Item:
         """Get item by id.
 
         Called with `GET /collections/{collection_id}/items/{item_id}`.
@@ -471,7 +471,7 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
         ...
 
     @abc.abstractmethod
-    def all_collections(self, **kwargs) -> stac_types.Collections:
+    def all_collections(self, **kwargs) -> api.Collections:
         """Get all available collections.
 
         Called with `GET /collections`.
@@ -482,7 +482,7 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
         ...
 
     @abc.abstractmethod
-    def get_collection(self, collection_id: str, **kwargs) -> stac_types.Collection:
+    def get_collection(self, collection_id: str, **kwargs) -> api.Collection:
         """Get collection by id.
 
         Called with `GET /collections/{collection_id}`.
@@ -499,12 +499,12 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
     def item_collection(
         self,
         collection_id: str,
-        bbox: Optional[List[NumType]] = None,
+        bbox: Optional[list[NumType]] = None,
         datetime: Optional[Union[str, datetime]] = None,
         limit: int = 10,
         token: str = None,
         **kwargs,
-    ) -> stac_types.ItemCollection:
+    ) -> api.ItemCollection:
         """Get all items from a specific collection.
 
         Called with `GET /collections/{collection_id}/items`
@@ -528,13 +528,13 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         extensions: list of registered api extensions.
     """
 
-    base_conformance_classes: List[str] = attr.ib(
+    base_conformance_classes: list[str] = attr.ib(
         factory=lambda: BASE_CONFORMANCE_CLASSES
     )
-    extensions: List[ApiExtension] = attr.ib(default=attr.Factory(list))
+    extensions: list[ApiExtension] = attr.ib(default=attr.Factory(list))
     post_request_model = attr.ib(default=BaseSearchPostRequest)
 
-    def conformance_classes(self) -> List[str]:
+    def conformance_classes(self) -> list[str]:
         """Generate conformance classes by adding extension conformance to base
         conformance classes."""
         conformance_classes = self.base_conformance_classes.copy()
@@ -549,7 +549,7 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         """Check if an api extension is enabled."""
         return any([type(ext).__name__ == extension for ext in self.extensions])
 
-    async def landing_page(self, **kwargs) -> stac_types.LandingPage:
+    async def landing_page(self, **kwargs) -> api.LandingPage:
         """Landing page.
 
         Called with `GET /`.
@@ -601,7 +601,7 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
 
         return landing_page
 
-    async def conformance(self, **kwargs) -> stac_types.Conformance:
+    async def conformance(self, **kwargs) -> api.ConformanceClasses:
         """Conformance classes.
 
         Called with `GET /conformance`.
@@ -609,12 +609,12 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         Returns:
             Conformance classes which the server conforms to.
         """
-        return Conformance(conformsTo=self.conformance_classes())
+        return api.ConformanceClasses(conformsTo=self.conformance_classes())
 
     @abc.abstractmethod
     async def post_search(
         self, search_request: BaseSearchPostRequest, **kwargs
-    ) -> stac_types.ItemCollection:
+    ) -> api.ItemCollection:
         """Cross catalog search (POST).
 
         Called with `POST /search`.
@@ -630,18 +630,18 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
     @abc.abstractmethod
     async def get_search(
         self,
-        collections: Optional[List[str]] = None,
-        ids: Optional[List[str]] = None,
-        bbox: Optional[List[NumType]] = None,
+        collections: Optional[list[str]] = None,
+        ids: Optional[list[str]] = None,
+        bbox: Optional[list[NumType]] = None,
         datetime: Optional[Union[str, datetime]] = None,
         limit: Optional[int] = 10,
         query: Optional[str] = None,
         token: Optional[str] = None,
-        fields: Optional[List[str]] = None,
+        fields: Optional[list[str]] = None,
         sortby: Optional[str] = None,
         intersects: Optional[str] = None,
         **kwargs,
-    ) -> stac_types.ItemCollection:
+    ) -> api.ItemCollection:
         """Cross catalog search (GET).
 
         Called with `GET /search`.
@@ -652,9 +652,7 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         ...
 
     @abc.abstractmethod
-    async def get_item(
-        self, item_id: str, collection_id: str, **kwargs
-    ) -> stac_types.Item:
+    async def get_item(self, item_id: str, collection_id: str, **kwargs) -> api.Item:
         """Get item by id.
 
         Called with `GET /collections/{collection_id}/items/{item_id}`.
@@ -669,7 +667,7 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         ...
 
     @abc.abstractmethod
-    async def all_collections(self, **kwargs) -> stac_types.Collections:
+    async def all_collections(self, **kwargs) -> api.Collections:
         """Get all available collections.
 
         Called with `GET /collections`.
@@ -680,9 +678,7 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         ...
 
     @abc.abstractmethod
-    async def get_collection(
-        self, collection_id: str, **kwargs
-    ) -> stac_types.Collection:
+    async def get_collection(self, collection_id: str, **kwargs) -> api.Collection:
         """Get collection by id.
 
         Called with `GET /collections/{collection_id}`.
@@ -699,12 +695,12 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
     async def item_collection(
         self,
         collection_id: str,
-        bbox: Optional[List[NumType]] = None,
+        bbox: Optional[list[NumType]] = None,
         datetime: Optional[Union[str, datetime]] = None,
         limit: int = 10,
         token: str = None,
         **kwargs,
-    ) -> stac_types.ItemCollection:
+    ) -> api.ItemCollection:
         """Get all items from a specific collection.
 
         Called with `GET /collections/{collection_id}/items`
@@ -726,7 +722,7 @@ class AsyncBaseFiltersClient(abc.ABC):
 
     async def get_queryables(
         self, collection_id: Optional[str] = None, **kwargs
-    ) -> Dict[str, Any]:
+    ) -> dict[str, Any]:
         """Get the queryables available for the given collection_id.
 
         If collection_id is None, returns the intersection of all queryables over all
@@ -752,7 +748,7 @@ class BaseFiltersClient(abc.ABC):
 
     def get_queryables(
         self, collection_id: Optional[str] = None, **kwargs
-    ) -> Dict[str, Any]:
+    ) -> dict[str, Any]:
         """Get the queryables available for the given collection_id.
 
         If collection_id is None, returns the intersection of all queryables over all